Health is lowered when Q is pressed. Healthbar shown on GUI is full if health is 1. Should probably figure out how to keep this proportional to the Health on the player which is currently 100. Or we can leave a full health on the healthbar as 1 and have the player have 1 health and go from there. Will try to add a different color for the healthbar

+++ Assets/Health.cs
@@ -2,9 +2,9 @@
 using System.Collections;
 
 public class Health : MonoBehaviour {
-	public int hp = 100;
+	public float hp = 100;
 
-	public void Damage(int damage) {
+	public void Damage(float damage) {
 		hp -= damage;
 
 		if (hp <= 0) {
